Style Header and main page
