+++ src/App.css
@@ -3,3 +3,4 @@
   background-position: center;
   background-size: cover;
 }
+
