+++ src/components/Header/Header.jsx
@@ -10,31 +10,29 @@ const Header = () => {
     navigate({ to: '/', replace: true });
   };
   return (
-    <header className="header py-4 px-24 shadow-xl flex items-center justify-between">
-      <NavLink to="/">
-        <h1 className="text-2xl font-bold">Microvago</h1>
+    <header className="max-h-screen py-6 shadow-xl text-center flex flex-col justify-between">
+      <NavLink to="/" className="text-2xl font-bold">
+        Microvago
       </NavLink>
-      <nav className="nav">
-        <ul className="navLinks flex gap-12 text-[gray] text-xl font-medium">
+      <nav className="nav row-span-3">
+        <ul className="row-span-3 navLinks flex flex-col gap-10 text-center text-[gray] text-xl font-medium">
           <li className="navLink hover:text-black">
-            <NavLink to="/addReservation">Add a reservation</NavLink>
+            <NavLink to="/addReservation">Add reservation</NavLink>
           </li>
           <li className="navLink hover:text-black">
             <NavLink to="/reservaions">My reservations</NavLink>
           </li>
           <li className="navLink hover:text-black">
-            <NavLink to="/addHotel">Add a Hotel</NavLink>
+            <NavLink to="/addHotel">Add Hotel</NavLink>
           </li>
           <li className="navLink hover:text-black">
-            <NavLink to="/delHotel">Delete a Hotel</NavLink>
-          </li>
-          <li className="navLink hover:text-black">
-            <button type="button" onClick={() => handleLogout()}>
-              Logout
-            </button>
+            <NavLink to="/delHotel">Delete Hotel</NavLink>
           </li>
         </ul>
       </nav>
+      <button type="button" onClick={() => handleLogout()}>
+        Logout
+      </button>
     </header>
   );
 };
